utils/code_extractor_simple.py
import cv2
import numpy as np
from PIL import Image
import os
import logging

logger = logging.getLogger(__name__)

def extract_codes_from_image(image_path: str) -> list:
    """
    Extract QR codes and barcodes from image (placeholder implementation)
    Returns empty list when pyzbar is not available
    """
    logger.warning("QR/barcode detection not available - pyzbar library missing")
    return []

def process_decoded_code(code) -> dict:
    """
    Placeholder function for processing decoded codes
    """
    return {
        'type': 'unknown',
        'data': '',
        'raw_data': '',
        'rect': {},
        'polygon': [],
        'extracted_info': {}
    }

def extract_code_info(code_data: str, code_type: str) -> dict:
    """
    Placeholder function for extracting code info
    """
    return {}

def extract_qr_info(qr_data: str) -> dict:
    """
    Placeholder function for QR info extraction
    """
    return {}

def extract_barcode_info(barcode_data: str, barcode_type: str) -> dict:
    """
    Placeholder function for barcode info extraction
    """
    return {}

def extract_code128_info(code_data: str) -> dict:
    """
    Placeholder function for Code 128 info extraction
    """
    return {}

def multi_scale_detection(image_path: str) -> list:
    """
    Placeholder function for multi-scale detection
    """
    return []

def enhance_for_code_detection(image_path: str) -> str:
    """
    Basic image enhancement without code detection
    """
    try:
        img = cv2.imread(image_path, cv2.IMREAD_GRAYSCALE)
        if img is None:
            return image_path
        
        # Apply histogram equalization
        enhanced = cv2.equalizeHist(img)
        
        # Create enhanced image path
        base_path, ext = os.path.splitext(image_path)
        enhanced_path = f"{base_path}_enhanced{ext}"
        
        cv2.imwrite(enhanced_path, enhanced)
        return enhanced_path
        
    except Exception as e:
        logger.error(f"Error enhancing image: {e}")
        return image_path